Fix shader with image.blend_mode = BlendMode.Premultiplied
Remove workaround removing fringing of png images when scrolling with sub-pixel precision. This workaround was causing incorrect blending on surfaces when glow was drawn on the surface. To fix fringing on scrolling png images with a mask it's advised to save png's with SuperPNG export plugin in photoshop with Clean Transparent option clicked, or in any other program by essentially making sure that fully transparent pixels are not set white during saving a file.